fix: fix a close bug

+++ script/dependency/con/run.py
@@ -75,9 +75,10 @@ class Process:
         p_cp_img = subprocess.Popen(cmd_cp_img, shell=True, preexec_fn=os.setsid)
         p_cp_img.wait()
 
-        cmd_cp_built_in = "cp ./built-in.* " + self.path
-        p_cp_img = subprocess.Popen(cmd_cp_built_in, shell=True, preexec_fn=os.setsid)
-        p_cp_img.wait()
+        if dra:
+            cmd_cp_built_in = "cp ./built-in.* " + self.path
+            p_cp_built_in = subprocess.Popen(cmd_cp_built_in, shell=True, preexec_fn=os.setsid)
+            p_cp_built_in.wait()
 
         f = open(os.path.join(path_root, file_json), "r")
         c = json.load(f)
@@ -117,21 +118,23 @@ class Process:
                        + file_bc + " 1>" + file_log_dra + " 2>&1"
         self.p_dra = subprocess.Popen(self.cmd_dra, shell=True, preexec_fn=os.setsid)
 
-    def close(self):
+    def close(self, dra=True):
         os.killpg(os.getpgid(self.p_syzkaller.pid), signal.SIGTERM)
-        os.killpg(os.getpgid(self.p_dra.pid), signal.SIGTERM)
+        if dra:
+            os.killpg(os.getpgid(self.p_dra.pid), signal.SIGTERM)
 
 
 def main():
+    dra = False
     tasks = [Process() for i in range(number_execute)]
     for i in tasks:
-        i.execute()
+        i.execute(dra)
 
-    time.sleep(time_run)
-    # time.sleep(30)
+    # time.sleep(time_run)
+    time.sleep(30)
 
     for i in tasks:
-        i.close()
+        i.close(dra)
 
 
 if __name__ == "__main__":
